handled screen rotations while doing background tasks
